DLS-11434: [SDIL] Apply Scala 3 Update to soft-drinks-industry-levy-returns-frontend

+++ app/repositories/SessionRepository.scala
@@ -44,7 +44,7 @@ class SessionRepository @Inject() (
         Indexes.ascending("lastUpdated"),
         IndexOptions()
           .name("lastUpdatedIdx")
-          .expireAfter(appConfig.cacheTtl, TimeUnit.SECONDS)))) {
+          .expireAfter(appConfig.cacheTtl.toLong, TimeUnit.SECONDS)))) {
 
   private def byId(id: String): Bson = Filters.equal("_id", id)
